Script extensions and code changes: support split-up lm-dir and creation of arpa file from split-up LM dir.  More testing in scripts

+++ scripts/make_lm_dir.py
@@ -6,10 +6,14 @@ import re, os, argparse, sys, math, warnings, subprocess, shutil
 from collections import defaultdict
 
 parser = argparse.ArgumentParser(description="This script, given counts and metaparameters, will "
-                                 "create an LM  in the 'pocolm-internal' format.  This consists of "
+                                 "estimate an LM  in the 'pocolm-internal' format.  This consists of "
                                  "a directory with a particular structure, containing the files: "
                                  "float.all (which contains most of the parameters), words.txt, "
-                                 "ngram_order, names, metaparameters and was_pruned.");
+                                 "ngram_order, names, metaparameters and was_pruned. "
+                                 "If you specify --keep-splits=true, the file num_splits will "
+                                 "be created (containing the number of split pieces of the counts), "
+                                 "and float.all won't exist, but float.all.split1, float.all.split2 "
+                                 "etc., will be created.");
 
 parser.add_argument('--fold-dev-into', type=str,
                     help='If supplied, the name of data-source into which to fold the '
@@ -19,6 +23,13 @@ parser.add_argument("--num-splits", type=int, default=1,
                     help="Controls the number of parallel processes used to "
                     "get objective functions and derivatives.  If >1, then "
                     "we split the counts and build the LM in parallel.")
+parser.add_argument("--keep-splits", type=str, choices=['true','false'],
+                    default='false',
+                    help="If true, instead of creating float.all, we'll create "
+                    "float.all.1, float.2 and so on, split by history-state according "
+                    "to the most recent history-word (the unigram state is repeated), "
+                    "and the file num_splits containing the --num-splits argument, which "
+                    "must be >1.")
 parser.add_argument("count_dir",
                     help="Directory from which to obtain counts files\n")
 parser.add_argument("metaparameters",
@@ -51,8 +62,19 @@ for name in [ 'ngram_order', 'num_train_sets', 'num_words' ]:
     f.close()
 
 
+if args.keep_splits == 'true':
+    if args.num_splits <= 1:
+        sys.exit("make_lm_dir.py: --num-splits must be >1 if "
+                 "you set --keep-splits=true")
+else:
+    if os.path.exists(args.lm_dir + "/num_splits"):
+        os.remove(args.lm_dir + "/num_splits")
+    if os.path.exists(args.lm_dir + "/float.all"):
+        os.remove(args.lm_dir + "/float.all")
+
+
 if args.num_splits < 1:
-    sys.exit("optimize_metaparameters.py: --num-splits must be >0.")
+    sys.exit("make_lm_dir.py: --num-splits must be >0.")
 if args.num_splits > 1:
     if (os.system("split_count_dir.sh {0} {1}".format(
                 args.count_dir, args.num_splits))) != 0:
@@ -109,12 +131,13 @@ if args.num_splits == 1:
                                                         metaparameters = args.metaparameters,
                                                         work_dir = work_dir))
 else:
-    command = ("get_objf_and_derivs_split.py --num-splits={num_splits} --need-model=true "
+    need_model_opt = '--need-model=true' if args.keep_splits == 'false' else ''
+    command = ("get_objf_and_derivs_split.py --num-splits={num_splits} {need_model_opt} "
                "{fold_dev_opt} {count_dir} {metaparameters} {work_dir}/objf "
                "{work_dir} 2>{work_dir}/log.txt".format(
-            fold_dev_opt = fold_dev_opt, num_splits = args.num_splits,
-            count_dir = args.count_dir, metaparameters = args.metaparameters,
-            work_dir = work_dir))
+            need_model_opt = need_model_opt, fold_dev_opt = fold_dev_opt,
+            num_splits = args.num_splits, count_dir = args.count_dir,
+            metaparameters = args.metaparameters, work_dir = work_dir))
 
 print("make_lm_dir.py: running command {0}".format(command), file=sys.stderr)
 
@@ -122,13 +145,27 @@ if os.system(command) != 0:
     sys.exit("make_lm_dir.py: error running command {0}".format(command))
 
 
-try:
-    shutil.move(work_dir + "/float.all",
-                args.lm_dir + "/float.all")
-except:
-    sys.exit("make_lm_dir.py: error moving {0}/float.all to {1}/float.all".format(
-            work_dir, args.lm_dir))
+if args.keep_splits == 'true':
+    f = open(args.lm_dir + '/num_splits', 'w')
+    print(str(args.num_splits), file=f)
+    f.close()
+    for i in range(1, args.num_splits + 1):
+        src_file = "{0}/split{1}/{2}/float.all".format(work_dir, args.num_splits, i)
+        dest_file = "{0}/float.all.{1}".format(args.lm_dir, i)
+        try:
+            shutil.move(src_file, dest_file)
+        except:
+            sys.exit("make_lm_dir.py: error moving {0} to {1}".format(src_file,
+                                                                      dest_file))
+else:
+    try:
+        shutil.move(work_dir + "/float.all",
+                    args.lm_dir + "/float.all")
+    except:
+        sys.exit("make_lm_dir.py: error moving {0}/float.all to {1}/float.all".format(
+                work_dir, args.lm_dir))
+
+
 
 if os.system("validate_lm_dir.py " + args.lm_dir) != 0:
     sys.exit("make_lm_dir.py: error validating lm-dir " + args.lm_dir)
-
